Add files via upload
This project was created as a project in one of the core courses named Fundamentals of Computer Programming(FCP). It is a project on Villa management in C language including the implementation of file management. One can log in as either Manager (and perform following functions staff management, revenue management, marketing management or sales management), Tariff (and can book a room available), place an order in Restaurant or can log in as a customer (and give ratings or register any complaint). All these have main C file named "Main Panel".